#2178 Extract BNF-to-text transformation into a reusable component
